(05-hf-batched-inference): wire Hydra config + W&B logging

+++ 05-hf-batched-inference/tests/conftest.py
@@ -3,4 +3,4 @@ from inference import loader, MODEL
 
 @pytest.fixture(scope="session")
 def model_and_tokenizer():
-    return loader(MODEL)
+    return loader(MODEL, dtype="bfloat16")
